+++ src/main/java/app/youngmon/surl/interfaces/UrlService.java
@@ -1,7 +1,6 @@
 package app.youngmon.surl.interfaces;
 
 public interface HashService {
-    char[]   getHashBaseArr();
     String   getShortUrl(String longUrl);
     String   getLongUrl(String shortUrl);
 }
